Create DataComponent class, based on StateManager
This means that we will be able to create a StateConsumer class without
duplicating similar code.

Had to update a type in MQTTWrapper that was too tightly coupled to
StateManager
